requirements.txt
# runtime: python 3.9.9

pandas
numpy
hydra-core